Extracts HasRelationships module
