+++ lib/userlist/rails/transform.rb
@@ -34,33 +34,9 @@ module Userlist
         (!model.respond_to?(:userlist_delete?) || model.userlist_delete?)
       end
 
-      def default_relationships
-        association = Userlist::Rails.find_association_between(relationship_from, relationship_to)
-
-        records = Array.wrap(model.try(association&.name))
-
-        if association.klass == config.relationship_model
-          records
-        else
-          records.map { |record| build_relationship(record) }
-        end
-      end
-
     private
 
       attr_reader :model, :config
-
-      def build_relationship(_record)
-        raise NotImplementedError
-      end
-
-      def relationship_to
-        raise NotImplementedError
-      end
-
-      def relationship_from
-        raise NotImplementedError
-      end
     end
   end
 end
